For simulations where selected sites have a distribution of effect sizes
For simulations where selected sites have a distribution of effect sizes. We had to make changes to how individuals are defined, namely that selected sites are explicitly defined. We also had to update the initialization procedures. Finally, we included an algorithm to estimate the effective population size based on the Hudson and Kaplan algorithm.